Remove close method
This method was used to call the finalizer.  This does not make
much sense since we can just use the finalize method.  Moreover,
it causes problems since once the finalizer is called, it gets
de-registered from the object so that subsequent calls are to close
are no-ops.  I will make a companion PR in JetPackWave that creates
custom close methods for each JetPackWave operator.